add more declarations
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@23520 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/libutils/bcmutils.c
@@ -1843,6 +1843,6 @@ int nvram_used(int *space)
 		name += strlen(name) + 1;
 	}
 
-	return (sizeof(struct nvram_header) + (int)name - (int)buf);
+	return (sizeof(struct nvram_header) + (long)name - (long)buf);
 
 }
